requirements.txt
python-dotenv
beautifulsoup4
requests
selenium
discord.py